requirements.txt
fastapi
uvicorn
python-dotenv
pyaudio
google-generativeai
websockets